feat(platform): provider detail drawer + api-key suffix masking
- Extract the 1.4k-line provider settings page into a reusable
  `provider-detail-drawer` component; the route now mounts the drawer.
- Add model picker UX: search, "show more" pagination, select-all,
  add-model confirmation, and a Showing-N-of-M counter.
- Mask provider API keys as `start … suffix` (6 chars + 4 chars) so
  users can match against the vendor dashboard. Adds a new
  `fetchConfiguredProviderModels` action that uses the stored secret.
- Capture the trailing 4 chars of Better Auth API keys at creation
  time via an after-hook on `/api-key/create`, store them in a new
  nullable `suffix` column on the `apikey` table, and render keys as
  `start … suffix` in the API Keys table. Pre-existing rows fall back
  to prefix-only.
- Drop the obsolete `examples/providers/openai.json` and shrink the
  bundled OpenRouter example.
- Translations for de / en / fr.
