+++ services/platform/app/features/settings/api-keys/components/api-key-revoke-dialog.test.tsx
@@ -20,6 +20,7 @@ function makeApiKey(overrides: Partial<ApiKey> = {}): ApiKey {
     name: 'Test Key',
     start: 'tale_abc',
     prefix: 'tale_',
+    suffix: 'wxyz',
     userId: 'user-1',
     enabled: true,
     expiresAt: null,
