+++ services/platform/app/features/settings/api-keys/hooks/use-api-keys-table-config.tsx
@@ -38,11 +38,15 @@ export function useApiKeysTableConfig(
       {
         id: 'key',
         header: tSettings('apiKeys.columns.key'),
-        cell: ({ row }) => (
-          <Text as="span" variant="muted" className="font-mono text-sm">
-            {row.original.start || row.original.prefix || '-'}
-          </Text>
-        ),
+        cell: ({ row }) => {
+          const head = row.original.start || row.original.prefix;
+          const tail = row.original.suffix;
+          return (
+            <Text as="span" variant="muted" className="font-mono text-sm">
+              {head ? (tail ? `${head} … ${tail}` : head) : '-'}
+            </Text>
+          );
+        },
       },
       {
         id: 'created',
