+++ services/platform/messages/en.json
@@ -1019,6 +1019,7 @@
       "editModel": "Edit model",
       "deleteModel": "Delete model",
       "deleteModelConfirm": "Are you sure you want to delete \"{model}\"?",
+      "addModelConfirm": "Add \"{model}\" to this provider? You can edit its details after adding.",
       "default": "default",
       "modelId": "Model ID",
       "displayName": "Display name",
@@ -1078,9 +1079,9 @@
       "save": "Save",
       "saving": "Saving...",
       "saveKey": "Save key",
-      "editGeneral": "Edit details",
+      "editGeneral": "Edit",
       "editGeneralTitle": "Edit general details",
-      "editDefaults": "Edit defaults",
+      "editDefaults": "Edit",
       "editDefaultModels": "Edit default models",
       "editKey": "Edit",
       "testConnection": "Test connection",
@@ -1113,13 +1114,16 @@
       "deleteConfirmNamed": "Are you sure you want to delete the provider \"{name}\"? This action cannot be undone. All model configurations associated with this provider will be removed.",
       "providerActions": "Provider actions",
       "saveChanges": "Save changes",
-      "addModelShort": "+ Add model",
+      "addModelShort": "Add model",
       "byomGuidanceTitle": "Getting started",
       "byomGuidance": "Any OpenAI-compatible endpoint works here. Point the base URL to your self-hosted inference server (e.g. vLLM, Ollama, LocalAI) or a third-party gateway like OpenRouter to use models such as LLaMA, Mistral, Gemini, and more.",
       "tagsRequired": "At least one tag is required",
       "modelsRequired": "At least one model is required",
       "fetchModels": "Fetch models",
+      "selectAllModels": "Select all models",
       "searchModels": "Search models…",
+      "showingModels": "Showing {filtered} of {total} models",
+      "showMoreModels": "Show more",
       "modelsEmpty": {
         "noModelsTitle": "No models yet",
         "noModelsDescription": "Add a model to make it available to this provider.",
@@ -1148,7 +1152,10 @@
       "tagImageEdit": "Image edit",
       "tagImageGeneration": "Image generation",
       "tagTextToSpeech": "Text-to-speech",
-      "tagTranscription": "Transcription"
+      "tagTranscription": "Transcription",
+      "details": "Provider details",
+      "capabilities": "Capabilities",
+      "apiKeyUpdated": "Your API key has been saved successfully."
     },
     "form": {
       "name": "Name",
