+++ services/platform/messages/fr.json
@@ -1019,6 +1019,7 @@
       "editModel": "Modifier le modèle",
       "deleteModel": "Supprimer le modèle",
       "deleteModelConfirm": "Es-tu sûr de vouloir supprimer « {model} » ?",
+      "addModelConfirm": "Ajouter « {model} » à ce fournisseur ? Tu pourras modifier les détails après l'ajout.",
       "default": "défaut",
       "modelId": "ID du modèle",
       "displayName": "Nom d'affichage",
@@ -1113,13 +1114,16 @@
       "deleteConfirmNamed": "Es-tu sûr de vouloir supprimer le fournisseur « {name} » ? Cette action est irréversible. Toutes les configurations de modèles associées à ce fournisseur seront supprimées.",
       "providerActions": "Actions du fournisseur",
       "saveChanges": "Enregistrer les modifications",
-      "addModelShort": "+ Ajouter un modèle",
+      "addModelShort": "Ajouter un modèle",
       "byomGuidanceTitle": "Pour commencer",
       "byomGuidance": "Tout point de terminaison compatible OpenAI fonctionne ici. Pointe l'URL de base vers ton serveur d'inférence auto-hébergé (ex. vLLM, Ollama, LocalAI) ou une passerelle tierce comme OpenRouter pour utiliser des modèles tels que LLaMA, Mistral, Gemini, et bien d'autres.",
       "tagsRequired": "Au moins un tag est requis",
       "modelsRequired": "Au moins un modèle est requis",
       "fetchModels": "Récupérer les modèles",
+      "selectAllModels": "Tout sélectionner",
       "searchModels": "Rechercher des modèles…",
+      "showingModels": "{filtered} sur {total} modèles",
+      "showMoreModels": "Afficher plus",
       "modelsEmpty": {
         "noModelsTitle": "Aucun modèle pour l'instant",
         "noModelsDescription": "Ajoute un modèle pour le rendre disponible pour ce fournisseur.",
@@ -1148,7 +1152,10 @@
       "tagImageEdit": "Édition d'images",
       "tagImageGeneration": "Génération d'images",
       "tagTextToSpeech": "Synthèse vocale",
-      "tagTranscription": "Transcription"
+      "tagTranscription": "Transcription",
+      "details": "Détails du fournisseur",
+      "capabilities": "Capacités",
+      "apiKeyUpdated": "Votre clé API a été enregistrée avec succès."
     },
     "form": {
       "name": "Nom",
